Refactor: 로직 분리 및 파일/함수명 수정, 파일/함수 이동
