+++ src/components/Conversation/Request/RequestChat.tsx
@@ -1,7 +1,7 @@
 import { Stack } from "@mui/material";
 import RequestMessageType from "../../../Interface/Message/RequestMessageType";
-import { useSmoothScrollToBottom } from "../../../utils/Chat";
-import RequestMessage from "../../Messages/RequestMessage";
+import { useSmoothScrollToBottom } from "../../../utils/chat";
+import RequestMessage from "../../Messages/Request/RequestMessage";
 
 export function RequestChat({ message }: { message: RequestMessageType }) {
 	const divRef = useSmoothScrollToBottom();
