+++ src/components/Messages/BasicMessage.tsx
@@ -1,0 +1,16 @@
+import { styled } from "@mui/material";
+
+export const BasicMessage = styled("div")`
+	z-index: 0;
+	padding: 8px 16px;
+	display: inline-block;
+
+	max-width: 70%;
+
+	background: #eaefef;
+	border-radius: 0px 20px 20px 20px;
+
+	color: black;
+	border: none;
+	box-shadow: none;
+`;
