+++ src/components/Messages/Request/RequestMessage.tsx
@@ -1,9 +1,10 @@
 import { Typography } from "@mui/material";
-import RequestMessageType from "../../Interface/Message/RequestMessageType";
+import RequestMessageType from "../../../Interface/Message/RequestMessageType";
 import MessageStyle from "./RequestMessageStyle";
 interface RequestMessageTypeProps {
 	message: RequestMessageType;
 }
+
 export default function RequestMessage({ message }: RequestMessageTypeProps) {
 	return (
 		<MessageStyle>
