+++ src/components/Messages/Response/ContentResponseMessage.tsx
@@ -1,26 +1,12 @@
 import { styled, Typography } from "@mui/material";
-import { ContentResponseMessageType } from "../../Interface/Message/ResponseMessageType";
-import { containsUrl } from "../../utils/Chat";
-import { primaryColor } from "../../utils/color";
+import { ContentResponseMessageType } from "../../../Interface/Message/ResponseMessageType";
+import { containsUrl } from "../../../utils/chat";
+import { primaryColor } from "../../../utils/color";
+import { BasicMessage } from "../BasicMessage";
 
 interface ContentResponseMessageTypeProps {
 	message: ContentResponseMessageType;
 }
-
-export const BasicMessage = styled("div")`
-	z-index: 0;
-	padding: 8px 16px;
-	display: inline-block;
-
-	max-width: 70%;
-
-	background: #eaefef;
-	border-radius: 0px 20px 20px 20px;
-
-	color: black;
-	border: none;
-	box-shadow: none;
-`;
 
 const StyledLink = styled("a")`
 	color: ${primaryColor};
