+++ src/components/Messages/Response/MessageButtons.tsx
@@ -1,45 +1,24 @@
-import styled from "@emotion/styled";
-import { Button, Stack } from "@mui/material";
+import { Stack } from "@mui/material";
 import { useRef } from "react";
 import {
 	ButtonResponseType,
 	imageResponseCardContentType,
-} from "../../Interface/Message/ResponseMessageType";
-import { primaryColor } from "../../utils/color";
-import { useAppDispatch, useAppSelector } from "../../store/store";
-import { fetchResponse } from "../../store/message/fetchResponse";
-import useMessageStatus from "../../hooks/Request/useMessageStatus";
+} from "../../../Interface/Message/ResponseMessageType";
+import useMessageStatus from "../../../hooks/Request/useMessageStatus";
+import { fetchResponse } from "../../../store/message/fetchResponse";
+import { useAppDispatch } from "../../../store/store";
+import { StyledButton } from "./StyledButton";
 
 interface MessageButtonsProps {
 	message: imageResponseCardContentType;
 	messageID: string;
 }
 
-interface IsSelectedInterface {
-	disabled: boolean;
-}
-
-const Style =
-	(trueProp: string, falseProp: string) => (props: IsSelectedInterface) => {
-		return props.disabled ? trueProp : falseProp;
-	};
-
-export const StyledButton = styled(Button)<IsSelectedInterface>`
-	padding-inline: 0.75rem;
-
-	background: ${Style("#eaefef", "white")};
-	color: ${Style(primaryColor, "black")};
-	text-align: start;
-
-	border-radius: 20px;
-	border: 1px solid ${Style(primaryColor, "#bed1d1")};
-	margin: 0 !important;
-`;
-
 export function MessageButtons({ message, messageID }: MessageButtonsProps) {
 	const clickedBtnListRef = useRef([] as string[]);
 	const dispatch = useAppDispatch();
 	const { status: isLoading } = useMessageStatus();
+
 	const sendRequest = (value: string) => {
 		dispatch(fetchResponse(value));
 	};
