+++ src/components/Messages/Response/StyledButton.tsx
@@ -1,0 +1,24 @@
+import styled from "@emotion/styled";
+import { Button } from "@mui/material";
+import { primaryColor } from "../../../utils/color";
+
+export interface IsSelectedInterface {
+	disabled: boolean;
+}
+
+export const Style =
+	(trueProp: string, falseProp: string) => (props: IsSelectedInterface) => {
+		return props.disabled ? trueProp : falseProp;
+	};
+
+export const StyledButton = styled(Button)<IsSelectedInterface>`
+	padding-inline: 0.75rem;
+
+	background: ${Style("#eaefef", "white")};
+	color: ${Style(primaryColor, "black")};
+	text-align: start;
+
+	border-radius: 20px;
+	border: 1px solid ${Style(primaryColor, "#bed1d1")};
+	margin: 0 !important;
+`;
